Move files for release candidate renaming. Update README.md for optimized conversion to rst.
