+++ openbadges/command_line.py
