+++ openbadges/verifier/__init__.py
@@ -1,0 +1,1 @@
+from .verifier import verify
