+++ openbadges/verifier/exceptions.py
