+++ openbadges/verifier/reducers/__init__.py
