+++ openbadges/verifier/reducers/graph.py
