+++ openbadges/verifier/reducers/input.py
