+++ openbadges/verifier/reducers/tasks.py
