+++ openbadges/verifier/reducers/verification_report.py
