+++ openbadges/verifier/state.py
