+++ openbadges/verifier/tasks/extensions.py
