+++ openbadges/verifier/tasks/images.py
