+++ openbadges/verifier/tasks/utils.py
