+++ openbadges/verifier/tasks/validation.py
