+++ openbadges/verifier/tasks/verification.py
