+++ openbadges/verifier/utils.py
