+++ openbadges/verifier/verifier.py
