+++ tests/test_validation.py
@@ -7,23 +7,23 @@ from pydux import create_store
 import responses
 import unittest
 
-from badgecheck.actions.action_types import ADD_TASK, PATCH_NODE
-from badgecheck.actions.graph import add_node, patch_node
-from badgecheck.actions.tasks import add_task
-from badgecheck.openbadges_context import OPENBADGES_CONTEXT_V2_DICT
-from badgecheck.reducers import main_reducer
-from badgecheck.state import filter_active_tasks, INITIAL_STATE
-from badgecheck.tasks import task_named
-from badgecheck.tasks.validation import (_get_validation_actions, assertion_timestamp_checks,
+from openbadges.verifier.actions.action_types import ADD_TASK, PATCH_NODE
+from openbadges.verifier.actions.graph import add_node, patch_node
+from openbadges.verifier.actions.tasks import add_task
+from openbadges.verifier.openbadges_context import OPENBADGES_CONTEXT_V2_DICT
+from openbadges.verifier.reducers import main_reducer
+from openbadges.verifier.state import filter_active_tasks, INITIAL_STATE
+from openbadges.verifier.tasks import task_named
+from openbadges.verifier.tasks.validation import (_get_validation_actions, assertion_timestamp_checks,
                                          criteria_property_dependencies, detect_and_validate_node_class,
                                          OBClasses, PrimitiveValueValidator, validate_property, ValueTypes,)
-from badgecheck.tasks.verification import (_default_verification_policy, hosted_id_in_verification_scope,)
-from badgecheck.tasks.task_types import (ASSERTION_TIMESTAMP_CHECKS, CRITERIA_PROPERTY_DEPENDENCIES,
+from openbadges.verifier.tasks.verification import (_default_verification_policy, hosted_id_in_verification_scope,)
+from openbadges.verifier.tasks.task_types import (ASSERTION_TIMESTAMP_CHECKS, CRITERIA_PROPERTY_DEPENDENCIES,
                                          DETECT_AND_VALIDATE_NODE_CLASS, HOSTED_ID_IN_VERIFICATION_SCOPE,
                                          IDENTITY_OBJECT_PROPERTY_DEPENDENCIES, ISSUER_PROPERTY_DEPENDENCIES,
                                          VALIDATE_RDF_TYPE_PROPERTY, VALIDATE_PROPERTY, VALIDATE_EXPECTED_NODE_CLASS)
-from badgecheck.utils import MESSAGE_LEVEL_WARNING
-from badgecheck.verifier import call_task, verify
+from openbadges.verifier.utils import MESSAGE_LEVEL_WARNING
+from openbadges.verifier.verifier import call_task, verify
 
 try:
     from .testfiles.test_components import test_components
