+++ tests/utils.py
@@ -1,7 +1,7 @@
 import os
 import responses
 
-from badgecheck.openbadges_context import OPENBADGES_CONTEXT_V2_URI
+from openbadges.verifier.openbadges_context import OPENBADGES_CONTEXT_V2_URI
 
 from .testfiles.test_components import test_components
